move lib.rs macros to macros.rs.
